stream_gen_python/requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
jmespath==1.0.0
protobuf==3.19.4
python-dateutil==2.8.2
python-dotenv==0.20.0
requests==2.27.1
s3transfer==0.5.2
six==1.16.0
urllib3==1.26.9
